fix signs issues
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@1605 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/sccp_device.h
@@ -60,7 +60,7 @@ int sccp_device_check_ringback(sccp_device_t * d);
 void * sccp_dev_postregistration(void *data);
 void sccp_dev_clean(sccp_device_t * d, boolean_t destroy, uint8_t cleanupTime);
 sccp_service_t * sccp_dev_serviceURL_find_byindex(sccp_device_t * d, uint8_t instance);
-int sccp_device_find_index_for_line(const sccp_device_t * d, char *lineName);
+uint16_t sccp_device_find_index_for_line(const sccp_device_t * d, char *lineName);
 
 void sccp_device_removeLine(sccp_device_t *device, sccp_line_t * l);
 int sccp_device_sendReset(sccp_device_t * d, uint8_t reset_type);
